class_assignments/04_match_finder/pyproject.toml
[project]
name = "04-match-finder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.0",
    "gspread>=6.2.1",
    "oauth2client>=4.1.3",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
